sdk: split unit and integration tests
