+++ sdk/state/integrationtests/helpers_test.go
@@ -1,4 +1,4 @@
-package integration
+package integrationtests
 
 import (
 	"crypto/rand"
@@ -17,7 +17,7 @@ import (
 
 // functions to be used in the state_test integration tests
 
-func initAccounts(t *testing.T, client horizonclient.ClientInterface, asset txnbuild.Asset, assetLimit string, distributorKP *keypair.Full) (initiator Participant, responder Participant) {
+func initAccounts(t *testing.T, asset txnbuild.Asset, assetLimit string, distributorKP *keypair.Full) (initiator Participant, responder Participant) {
 	initiator = Participant{
 		Name:         "Initiator",
 		KP:           keypair.MustRandom(),
@@ -27,11 +27,11 @@ func initAccounts(t *testing.T, client horizonclient.ClientInterface, asset txnb
 	t.Log("Initiator:", initiator.KP.Address())
 	t.Log("Initiator Escrow:", initiator.Escrow.Address())
 	{
-		err := retry(2, func() error { return createAccount(client, initiator.KP.FromAddress(), 10_000_0000000) })
+		err := retry(2, func() error { return createAccount(initiator.KP.FromAddress(), 10_000_0000000) })
 		require.NoError(t, err)
-		err = retry(2, func() error { return fundAsset(client, asset, initiator.Contribution, initiator.KP, distributorKP) })
+		err = retry(2, func() error { return fundAsset(asset, initiator.Contribution, initiator.KP, distributorKP) })
 		require.NoError(t, err)
-		initEscrowAccount(t, client, &initiator, asset, assetLimit)
+		initEscrowAccount(t, &initiator, asset, assetLimit)
 	}
 
 	t.Log("Initiator Escrow Sequence Number:", initiator.EscrowSequenceNumber)
@@ -47,18 +47,18 @@ func initAccounts(t *testing.T, client horizonclient.ClientInterface, asset txnb
 	t.Log("Responder:", responder.KP.Address())
 	t.Log("Responder Escrow:", responder.Escrow.Address())
 	{
-		err := retry(2, func() error { return createAccount(client, responder.KP.FromAddress(), 10_000_0000000) })
+		err := retry(2, func() error { return createAccount(responder.KP.FromAddress(), 10_000_0000000) })
 		require.NoError(t, err)
-		err = retry(2, func() error { return fundAsset(client, asset, responder.Contribution, responder.KP, distributorKP) })
+		err = retry(2, func() error { return fundAsset(asset, responder.Contribution, responder.KP, distributorKP) })
 		require.NoError(t, err)
-		initEscrowAccount(t, client, &responder, asset, assetLimit)
+		initEscrowAccount(t, &responder, asset, assetLimit)
 	}
 	t.Log("Responder Escrow Sequence Number:", responder.EscrowSequenceNumber)
 	t.Log("Responder Contribution:", responder.Contribution, "of asset:", asset.GetCode(), "issuer: ", asset.GetIssuer())
 	return initiator, responder
 }
 
-func initEscrowAccount(t *testing.T, client horizonclient.ClientInterface, participant *Participant, asset txnbuild.Asset, assetLimit string) {
+func initEscrowAccount(t *testing.T, participant *Participant, asset txnbuild.Asset, assetLimit string) {
 	// create escrow account
 	account, err := client.AccountDetail(horizonclient.AccountRequest{AccountID: participant.KP.Address()})
 	require.NoError(t, err)
@@ -111,7 +111,7 @@ func initEscrowAccount(t *testing.T, client horizonclient.ClientInterface, parti
 	require.NoError(t, err)
 }
 
-func initChannels(t *testing.T, client horizonclient.ClientInterface, initiator Participant, responder Participant) (initiatorChannel *state.Channel, responderChannel *state.Channel) {
+func initChannels(t *testing.T, initiator Participant, responder Participant) (initiatorChannel *state.Channel, responderChannel *state.Channel) {
 	// Channel constants.
 	const observationPeriodTime = 20 * time.Second
 	const averageLedgerDuration = 5 * time.Second
@@ -153,9 +153,9 @@ func initAsset(t *testing.T, client horizonclient.ClientInterface) (txnbuild.Ass
 	issuerKP := keypair.MustRandom()
 	distributorKP := keypair.MustRandom()
 
-	err := retry(2, func() error { return createAccount(client, issuerKP.FromAddress(), 1_000_0000000) })
+	err := retry(2, func() error { return createAccount(issuerKP.FromAddress(), 1_000_0000000) })
 	require.NoError(t, err)
-	err = retry(2, func() error { return createAccount(client, distributorKP.FromAddress(), 1_000_0000000) })
+	err = retry(2, func() error { return createAccount(distributorKP.FromAddress(), 1_000_0000000) })
 	require.NoError(t, err)
 
 	distributor, err := client.AccountDetail(horizonclient.AccountRequest{AccountID: distributorKP.Address()})
@@ -218,7 +218,7 @@ func retry(maxAttempts int, f func() error) (err error) {
 	return err
 }
 
-func fundAsset(client horizonclient.ClientInterface, asset txnbuild.Asset, amount int64, accountKP *keypair.Full, distributorKP *keypair.Full) error {
+func fundAsset(asset txnbuild.Asset, amount int64, accountKP *keypair.Full, distributorKP *keypair.Full) error {
 	distributor, err := client.AccountDetail(horizonclient.AccountRequest{AccountID: distributorKP.Address()})
 	if err != nil {
 		return err
@@ -264,7 +264,7 @@ func fundAsset(client horizonclient.ClientInterface, asset txnbuild.Asset, amoun
 	return nil
 }
 
-func createAccount(client horizonclient.ClientInterface, account *keypair.FromAddress, startingBalance int64) error {
+func createAccount(account *keypair.FromAddress, startingBalance int64) error {
 	rootResp, err := client.Root()
 	if err != nil {
 		return err
